third_party/WebKit/LayoutTests/fast/speech/input-text-speechstart-expected.txt

CONSOLE WARNING: The 'x-webkit-speech' input field attribute is deprecated. Please use the JavaScript API instead.
Tests for triggering speech directly in a <input type="text" speech>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("speechInput").value is "Pictures of a sunset"
PASS successfullyParsed is true

TEST COMPLETE
